HEHE/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
pypdf
python-docx